+++ src/auth/auth.controller.spec.ts
@@ -4,13 +4,21 @@ import { AuthService } from './auth.service';
 import { CredentialsModule } from '@app/credentials/credentials.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule } from '@nestjs/config';
+import mockConfig from '@app/config/mock.config';
 
 describe('AuthController', () => {
   let controller: AuthController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [CredentialsModule, JwtModule, ConfigModule],
+      imports: [
+        CredentialsModule,
+        JwtModule,
+        ConfigModule.forRoot({
+          isGlobal: true,
+          load: [mockConfig],
+        }),
+      ],
       controllers: [AuthController],
       providers: [AuthService],
     }).compile();