+++ src/auth/auth.guard.ts
@@ -1,3 +1,4 @@
+import { AuthenticatedRequest } from '@app/interfaces/authenticated_request.auth.interface';
 import {
   CanActivate,
   ExecutionContext,
@@ -6,7 +7,7 @@ import {
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
-import { Request } from 'express';
+import { Request as ExpressRequest } from 'express';
 
 // Imported from NestJS documentation: https://docs.nestjs.com/security/authentication
 @Injectable()
@@ -17,28 +18,29 @@ export class AuthGuard implements CanActivate {
   ) {}
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
-    const request = context.switchToHttp().getRequest<Request>();
+    const request = context.switchToHttp().getRequest<AuthenticatedRequest>();
     const token = this.extractTokenFromHeader(request);
     if (!token) {
       throw new UnauthorizedException();
     }
     try {
-      const payload = await this.jwtService.verifyAsync<{ username: string }>(
-        token,
-        {
-          secret: this.config.get<string>('auth.secret_key'),
-        },
-      );
+      const payload = await this.jwtService.verifyAsync<{
+        sub: string;
+        email: string;
+      }>(token, {
+        secret: this.config.get<string>('auth.secret_key'),
+      });
       // 💡 We're assigning the payload to the request object here
       // so that we can access it in our route handlers
-      request['credentials'] = payload;
+      request.email = payload.email;
+      request.userId = payload.sub;
     } catch {
       throw new UnauthorizedException();
     }
     return true;
   }
 
-  private extractTokenFromHeader(request: Request): string | undefined {
+  private extractTokenFromHeader(request: ExpressRequest): string | undefined {
     const [type, token] = request.headers.authorization?.split(' ') ?? [];
     return type === 'Bearer' ? token : undefined;
   }